docs: commenting the code
