+++ pages/index.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
-import { onMounted, watch, ref, computed } from 'vue';
-import { useWebSocketSubscription } from '~/composables/useWebSocketSubscription';
+import { useBlockFeed } from '~/composables/useBlockFeed';
 
 definePageMeta({
   layout: 'app',
@@ -40,131 +39,7 @@ const { data: cgData, status: cgStatus, error: cgError } = await useAsyncData('h
   lazy: true,
 });
 
-const displayedBlockGroups = ref(new Map());
-
-const sortedBlockGroups = computed(() => {
-  return Array.from(displayedBlockGroups.value.values()).sort((a, b) => b.height - a.height);
-});
-
-function updateTotalTransactions(blockGroup: any) {
-  blockGroup.totalTransactions = Array.from(blockGroup.chains.values()).reduce((sum, chain: any) => sum + (chain.transactions?.totalCount || 0), 0);
-}
-
-const { startSubscription, newBlocks } = useWebSocketSubscription();
-
-const completedBlockHeightsQuery = `
-  query CompletedBlockHeights($heightCount: Int, $completedHeights: Boolean, $first: Int) {
-    completedBlockHeights(heightCount: $heightCount, completedHeights: $completedHeights, first: $first) {
-      edges {
-        node {
-          chainId
-          creationTime
-          height
-          transactions {
-            totalCount
-          }
-        }
-      }
-    }
-  }
-`;
-
-const fetchInitialBlocks = async () => {
-  const GRAPHQL_URL = 'https://mainnet.kadindexer.io/graphql';
-
-  try {
-    const response = await $fetch(GRAPHQL_URL, {
-      method: 'POST',
-      body: {
-        query: completedBlockHeightsQuery,
-        variables: {
-          heightCount: 6,
-          completedHeights: false,
-          first: 120, // 6 heights * 20 chains = 120
-        },
-      },
-    });
-
-    const responseData = response?.data?.completedBlockHeights;
-
-    if (responseData && responseData.edges) {
-      responseData.edges.forEach((edge) => {
-        const block = edge.node;
-        let blockToUpdate = displayedBlockGroups.value.get(block.height);
-
-        if (blockToUpdate) {
-          blockToUpdate.chains.set(block.chainId, block);
-          updateTotalTransactions(blockToUpdate);
-        } else {
-          const newGroup = {
-            height: block.height,
-            chains: new Map([[block.chainId, block]]),
-            createdAt: block.creationTime,
-            totalTransactions: block.transactions?.totalCount || 0,
-          };
-          displayedBlockGroups.value.set(block.height, newGroup);
-        }
-      });
-
-      const sortedGroups = Array.from(displayedBlockGroups.value.values()).sort((a, b) => b.height - a.height);
-      if (sortedGroups.length > 6) {
-        const cutoffHeight = sortedGroups[5].height;
-        for (const height of displayedBlockGroups.value.keys()) {
-          if (height < cutoffHeight) {
-            displayedBlockGroups.value.delete(height);
-          }
-        }
-      }
-    }
-  } catch (e) {
-    console.error('Failed to fetch initial blocks:', e);
-  }
-};
-
-onMounted(async () => {
-  await fetchInitialBlocks();
-  startSubscription();
-});
-
-watch(newBlocks, (latestBlocks) => {
-  latestBlocks.forEach((block: any) => {
-    let blockToUpdate = displayedBlockGroups.value.get(block.height);
-
-    if (blockToUpdate) {
-      blockToUpdate.chains.set(block.chainId, block);
-      updateTotalTransactions(blockToUpdate);
-    } else {
-      const heights = Array.from(displayedBlockGroups.value.keys());
-      if (heights.length === 0) { // First block ever
-        const newGroup = {
-          height: block.height,
-          chains: new Map([[block.chainId, block]]),
-          createdAt: block.creationTime,
-          totalTransactions: block.transactions?.totalCount || 0,
-        };
-        displayedBlockGroups.value.set(block.height, newGroup);
-        return;
-      }
-
-      const minHeight = Math.min(...heights);
-
-      if (block.height > minHeight || displayedBlockGroups.value.size < 6) {
-        const newGroup = {
-          height: block.height,
-          chains: new Map([[block.chainId, block]]),
-          createdAt: block.creationTime,
-          totalTransactions: block.transactions?.totalCount || 0,
-        };
-        displayedBlockGroups.value.set(block.height, newGroup);
-
-        if (displayedBlockGroups.value.size > 6) {
-            const sortedHeights = Array.from(displayedBlockGroups.value.keys()).sort((a,b) => a-b);
-            displayedBlockGroups.value.delete(sortedHeights[0]);
-        }
-      }
-    }
-  });
-}, { deep: true });
+const { sortedBlockGroups } = useBlockFeed();
 </script>
 
 <template>
